Fix AdoptionCommand when only kwin-x11 is installed
We are in an awkward situation where there are several sources of decoration
knsrc files: the aurorae engine itself and the compositor. We missed the
chance to sort it out during Plasma 6 transition so we must still keep it.

Both kwin_wayland and kwin_x11 provide their own adoption helpers, and
aurorae needs to use either one of them, they are compatible with each other.

This change adds a tiny wrapper that is going to invoke either
kwin-applywindowdecoration or kwin-applywindowdecoration-x11 to make sure
that decorations installed using aurorae.knsrc operate as expected.